% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitStatistics.R
\name{FitStatistics}
\alias{FitStatistics}
\alias{FitStatistics,ANY-method}
\title{Calculating fit statistics}
\usage{
FitStatistics(y, p, r = NULL, fits)
}
\arguments{
\item{y}{A numeric vector with observations}

\item{p}{A matrix of predicted values}

\item{r}{A numeric vector of fit statistics: RMSE, MAD, RMSLE, MAPE, MEAPE}
}
\value{
A matrix of fit statistic results of each model
}
\description{
Evaluate the fit of various statistical models
}
\note{
Fit statistics include RMSE(root mean squared error), 
MAD(median absolute deviation), MEAPE(median absolute percentage error), RMSLE, and MAPE
}
\examples{

myX <- rnorm(100, 0, 1) 
myY <- sample(size = 100, seq(1:1000))
p <- matrix(predict(lm(myY ~ myX), ncol=1))
r <- sample(min(myY):max(myY), 100, replace = TRUE)
FitStatistics(y=myY, p=p, r=r, fits=c("rmse", "mad", "meape"))
}
\author{
Hyunjoo Oh<\email{hyunjoo.oh@wustl.edu}>
}

